updated changelog and docker

+++ Docker/__test_run_Docker.sh
@@ -2,17 +2,17 @@
 
 set -ve
 
-CTAT_GENOME_LIB="GRCh37_v19_CTAT_lib_Feb092018"
+CTAT_GENOME_LIB="GRCh38_gencode_v29_CTAT_lib_Mar272019.plug-n-play"
 
-CTAT_GENOME_LIB_URL="https://data.broadinstitute.org/Trinity/CTAT_RESOURCE_LIB/__genome_libs_StarFv1.3/GRCh37_v19_CTAT_lib_Feb092018.plug-n-play.tar.gz"
+CTAT_GENOME_LIB_URL="https://data.broadinstitute.org/Trinity/CTAT_RESOURCE_LIB/__genome_libs_StarFv1.6/GRCh38_gencode_v29_CTAT_lib_Mar272019.plug-n-play.tar.gz"
 
 
-if [ ! -s "../${CTAT_GENOME_LIB}.tar.gz" ]; then
+if [ ! -s "../${CTAT_GENOME_LIB}.tar.gz" ] && [ ! -d "../${CTAT_GENOME_LIB}" ]; then
     wget ${CTAT_GENOME_LIB_URL} -O ../${CTAT_GENOME_LIB}.tar.gz
 fi
 
 
-if [ ! -d "../${CTAT_GENOME_LIB}" ]; then
+if [ ! -d "../${CTAT_GENOME_LIB}/" ]; then
     tar xvf ../${CTAT_GENOME_LIB}.tar.gz -C ../.
 fi
